// fe/EsercitazioneFinaleAngular2/src/enums/api-paths.ts
export enum ApiPaths {

    Main = "http://localhost:8080",

    Rooms = "/rooms",
    Bookings = "/bookings",
    Sites = "/sites",



    SiteRooms = "/siteRooms",

    roomBookings = "/roomBookings",

    bookingById = "/bookingById",

    deleteBooking = "/deleteBooking",

    roomById = "/roomById",



    bookingByRoomIdOrderByDate = "/bookingByRoomIdOrderByDate",


    userByEmail = "/userByEmail",

    checkUser = "/checkUser",

    exportCsv = "/csvexport"
}